+++ src/settings/PreferencesSettingsTab.tsx
@@ -14,6 +14,7 @@ import {
   showHandRaisedTimer as showHandRaisedTimerSetting,
   showReactions as showReactionsSetting,
   playReactionsSound as playReactionsSoundSetting,
+  developerSettingsTab as developerSettingsTabSetting,
   useSetting,
 } from "./settings";
 
@@ -36,9 +37,12 @@ export const PreferencesSettingsTab: FC = () => {
     fn(e.target.checked);
   };
 
+  const [developerSettingsTab, setDeveloperSettingsTab] = useSetting(
+    developerSettingsTabSetting,
+  );
+
   return (
     <div>
-      <h4>{t("settings.preferences_tab_h4")}</h4>
       <Text>{t("settings.preferences_tab_body")}</Text>
       <FieldRow>
         <InputField
@@ -52,7 +56,6 @@ export const PreferencesSettingsTab: FC = () => {
           onChange={(e) => onChangeSetting(e, setShowHandRaisedTimer)}
         />
       </FieldRow>
-      <h5>{t("settings.preferences_tab.reactions_title")}</h5>
       <FieldRow>
         <InputField
           id="showReactions"
@@ -73,6 +76,20 @@ export const PreferencesSettingsTab: FC = () => {
           type="checkbox"
           checked={playReactionsSound}
           onChange={(e) => onChangeSetting(e, setPlayReactionSound)}
+        />
+      </FieldRow>
+      <FieldRow>
+        <InputField
+          id="developerSettingsTab"
+          type="checkbox"
+          checked={developerSettingsTab}
+          label={t("settings.preferences_tab.developer_mode_label")}
+          description={t(
+            "settings.preferences_tab.developer_mode_label_description",
+          )}
+          onChange={(event: ChangeEvent<HTMLInputElement>): void =>
+            setDeveloperSettingsTab(event.target.checked)
+          }
         />
       </FieldRow>
     </div>
